requirements.txt
# ──────────────────────────────────────────────
# AI Agent Microservice — Python Dependencies
# ──────────────────────────────────────────────

# --- Web / REST ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- gRPC ---
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2

# --- Async HTTP client (Ollama) ---
httpx==0.28.1

# --- Settings & Validation ---
pydantic==2.10.4
pydantic-settings==2.7.1

# --- Observability ---
mlflow==2.19.0

# --- Logging ---
structlog==24.4.0

# --- Utilities ---
tenacity==9.0.0
python-dotenv==1.0.1
orjson==3.10.13
